yt_service/requirements.service.txt
Django==2.0.2
django-extensions==1.9.8
djangorestframework==3.7.3
pytest==3.3.0
djangorestframework==3.7.3
six==1.11.0
djangorestframework-jwt==1.11.0
requests==2.18.4
django-cors-headers
pandas==1.3.4
numpy==1.20.3
neurokit2==0.1.5